requirements.txt
aiogram
google-generativeai
requests